--- relapse/funcs/simplify.go
@@ -1,43 +1,0 @@
-//  Copyright 2016 Walter Schulze
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-
-package funcs
-
-//IsFalse returns whether a function is a false constant.
-func IsFalse(fn Bool) bool {
-	v, ok := fn.(*constBool)
-	if !ok {
-		return false
-	}
-	return v.v == false
-}
-
-//IsTrue returns whether a function is a true constant.
-func IsTrue(fn Bool) bool {
-	v, ok := fn.(*constBool)
-	if !ok {
-		return false
-	}
-	return v.v == true
-}
-
-//Equal returns whether two functions are equal.
-func Equal(l, r Comparable) bool {
-	hl := l.Hash()
-	hr := r.Hash()
-	if hl != hr {
-		return false
-	}
-	return l.Compare(r) == 0
-}
